+++ supertokens/constants.go
@@ -21,10 +21,10 @@ const (
 )
 
 // VERSION current version of the lib
-const VERSION = "0.23.0"
+const VERSION = "0.24.0"
 
 var (
-	cdiSupported = []string{"3.0"}
+	cdiSupported = []string{"3.1"}
 )
 
 const DashboardVersion = "0.7"