+++ supertokens/querier.go
@@ -22,6 +22,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"net/url"
 	"sort"
 	"strings"
 	"sync"
@@ -45,7 +46,7 @@ var (
 	querierLastTriedIndex int
 	querierLock           sync.Mutex
 	querierHostLock       sync.Mutex
-	querierInterceptor    func(*http.Request, UserContext) *http.Request
+	querierInterceptor    func(*http.Request, UserContext) (*http.Request, error)
 	querierGlobalCacheTag uint64
 	querierDisableCache   bool
 )
@@ -54,20 +55,64 @@ func SetQuerierApiVersionForTests(version string) {
 	querierAPIVersion = version
 }
 
-func (q *Querier) GetQuerierAPIVersion() (string, error) {
+func (q *Querier) GetQuerierAPIVersion(userContextIn ...UserContext) (string, error) {
 	querierLock.Lock()
 	defer querierLock.Unlock()
 	if querierAPIVersion != "" {
 		return querierAPIVersion, nil
 	}
+
+	var userContext UserContext = nil
+	if len(userContextIn) > 0 {
+		userContext = userContextIn[0]
+	}
+
+	appInfo := superTokensInstance.AppInfo
+	req := getRequestFromUserContext(userContext)
+	websiteDomain, err := appInfo.GetOrigin(req, userContext)
+	if err != nil {
+		return "", err
+	}
+	queryParamsObj := map[string]string{
+		"apiDomain":     appInfo.APIDomain.GetAsStringDangerous(),
+		"websiteDomain": websiteDomain.GetAsStringDangerous(),
+	}
+
+	var queryParams []string
+	for key, value := range queryParamsObj {
+		queryParams = append(queryParams, fmt.Sprintf("%s=%s", key, url.QueryEscape(value)))
+	}
+	queryString := strings.Join(queryParams, "&")
+
 	response, _, err := q.sendRequestHelper(NormalisedURLPath{value: "/apiversion"}, func(url string) (*http.Response, []byte, error) {
 		req, err := http.NewRequest("GET", url, nil)
 		if err != nil {
 			return nil, nil, err
 		}
+
+		headers := make(http.Header)
 		if QuerierAPIKey != nil {
-			req.Header.Set("api-key", *QuerierAPIKey)
+			headers.Set("api-key", *QuerierAPIKey)
 		}
+
+		// Apply network interceptor if available
+		if querierInterceptor != nil {
+			interceptedReq := &http.Request{
+				URL:    req.URL,
+				Method: req.Method,
+				Header: headers,
+			}
+			interceptedReq.URL.RawQuery = queryString
+			interceptedReq, err = querierInterceptor(interceptedReq, userContext)
+			if err != nil {
+				return nil, nil, err
+			}
+			req.URL = interceptedReq.URL
+			req.Header = interceptedReq.Header
+		} else {
+			req.Header = headers
+		}
+
 		client := &http.Client{}
 		resp, err := client.Do(req)
 		return resp, nil, err
@@ -105,7 +150,7 @@ func GetNewQuerierInstanceOrThrowError(rIDToCore string) (*Querier, error) {
 	return &Querier{RIDToCore: rIDToCore}, nil
 }
 
-func initQuerier(hosts []QuerierHost, APIKey string, interceptor func(*http.Request, UserContext) *http.Request, disableCache bool) {
+func initQuerier(hosts []QuerierHost, APIKey string, interceptor func(*http.Request, UserContext) (*http.Request, error), disableCache bool) {
 	if !querierInitCalled {
 		querierInitCalled = true
 		QuerierHosts = hosts
@@ -139,7 +184,7 @@ func (q *Querier) SendPostRequest(path string, data map[string]interface{}, user
 			return nil, nil, err
 		}
 
-		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion()
+		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion(userContext)
 		if querierAPIVersionError != nil {
 			return nil, nil, querierAPIVersionError
 		}
@@ -154,7 +199,10 @@ func (q *Querier) SendPostRequest(path string, data map[string]interface{}, user
 		}
 
 		if querierInterceptor != nil {
-			req = querierInterceptor(req, userContext)
+			req, err = querierInterceptor(req, userContext)
+			if err != nil {
+				return nil, nil, err
+			}
 		}
 
 		client := &http.Client{}
@@ -187,7 +235,7 @@ func (q *Querier) SendDeleteRequest(path string, data map[string]interface{}, pa
 		}
 		req.URL.RawQuery = query.Encode()
 
-		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion()
+		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion(userContext)
 		if querierAPIVersionError != nil {
 			return nil, nil, querierAPIVersionError
 		}
@@ -202,7 +250,10 @@ func (q *Querier) SendDeleteRequest(path string, data map[string]interface{}, pa
 		}
 
 		if querierInterceptor != nil {
-			req = querierInterceptor(req, userContext)
+			req, err = querierInterceptor(req, userContext)
+			if err != nil {
+				return nil, nil, err
+			}
 		}
 
 		client := &http.Client{}
@@ -247,7 +298,7 @@ func (q *Querier) SendGetRequest(path string, params map[string]string, userCont
 		// Append sorted headers to the unique key
 		headers := make(map[string]string)
 
-		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion()
+		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion(userContext)
 		if querierAPIVersionError != nil {
 			return nil, nil, querierAPIVersionError
 		}
@@ -303,7 +354,10 @@ func (q *Querier) SendGetRequest(path string, params map[string]string, userCont
 		}
 
 		if querierInterceptor != nil {
-			req = querierInterceptor(req, userContext)
+			req, err = querierInterceptor(req, userContext)
+			if err != nil {
+				return nil, nil, err
+			}
 		}
 
 		client := &http.Client{}
@@ -362,7 +416,7 @@ func (q *Querier) SendGetRequestWithResponseHeaders(path string, params map[stri
 		}
 		req.URL.RawQuery = query.Encode()
 
-		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion()
+		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion(userContext)
 		if querierAPIVersionError != nil {
 			return nil, nil, querierAPIVersionError
 		}
@@ -375,7 +429,10 @@ func (q *Querier) SendGetRequestWithResponseHeaders(path string, params map[stri
 		}
 
 		if querierInterceptor != nil {
-			req = querierInterceptor(req, userContext)
+			req, err = querierInterceptor(req, userContext)
+			if err != nil {
+				return nil, nil, err
+			}
 		}
 
 		client := &http.Client{}
@@ -400,7 +457,7 @@ func (q *Querier) SendPutRequest(path string, data map[string]interface{}, userC
 			return nil, nil, err
 		}
 
-		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion()
+		apiVersion, querierAPIVersionError := q.GetQuerierAPIVersion(userContext)
 		if querierAPIVersionError != nil {
 			return nil, nil, querierAPIVersionError
 		}
@@ -415,7 +472,10 @@ func (q *Querier) SendPutRequest(path string, data map[string]interface{}, userC
 		}
 
 		if querierInterceptor != nil {
-			req = querierInterceptor(req, userContext)
+			req, err = querierInterceptor(req, userContext)
+			if err != nil {
+				return nil, nil, err
+			}
 		}
 
 		client := &http.Client{}